requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy
opencv-python
scikit-image
tqdm
timm
einops
natten
lxml
Pillow
wandb
